application/src/test/java/uk/nhs/hee/tis/revalidation/messages/publisher/GmcDoctorsForDbSyncStartPublisherTest.java
/*
 * The MIT License (MIT)
 *
 * Copyright 2021 Crown Copyright (Health Education England)
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and
 * associated documentation files (the "Software"), to deal in the Software without restriction,
 * including without limitation the rights to use, copy, modify, merge, publish, distribute,
 * sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is
 * furnished to do so, subject to the following conditions:
 *
 * The above copyright notice and this permission notice shall be included in all copies or
 * substantial portions of the Software.
 *
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT
 * NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
 * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
 * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */

package uk.nhs.hee.tis.revalidation.messages.publisher;

import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.ArgumentCaptor;
import org.mockito.Captor;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;
import org.springframework.test.util.ReflectionTestUtils;

import static org.hamcrest.CoreMatchers.is;
import static org.hamcrest.MatcherAssert.assertThat;
import static org.mockito.Mockito.verify;

@ExtendWith(MockitoExtension.class)
public class GmcDoctorsForDbSyncStartPublisherTest {

  @InjectMocks
  GmcDoctorsForDbSyncStartPublisher gmcDoctorsForDbSyncStartPublisher;

  @Mock
  MessagePublisher<String> messagePublisher;

  @Captor
  ArgumentCaptor<String> messageCaptor;

  @Test
  public void shouldPublishStartSyncMessage() {

    final String startMessage = "start";
    ReflectionTestUtils.setField(
        gmcDoctorsForDbSyncStartPublisher, "startMessage", startMessage
    );

    gmcDoctorsForDbSyncStartPublisher.publishNightlySyncStartMessage();

    verify(messagePublisher).publishToBroker(
        messageCaptor.capture()
    );

    assertThat(messageCaptor.getValue(), is(startMessage));

  }


}
